requirements.txt
aiogram>=3.20.0
httpx>=0.27.0
python-dotenv>=1.0.1
matplotlib>=3.8
